Created AVLTree
Created a node class which has fields for value in node, left node
child, right node child, and height. This class also has constructor
method to set the value and a toString method to display value of node
while debugging.

Also created insert method which adds values recursively, while also
calculating height recursively with the use of a helper height method
which has ternary operator to check if a child node is null.